@@ -1,31 +1,31 @@
 <template lang="pug">
-  #footerwrap
-    .container
-      .row.centered
-        .col-lg-12
-          h2
-            sup
-              i.fa.fa-quote-left
-            | {{ quote }}
-            sup
-              i.fa.fa-quote-right
-        .col-lg-4
-          p
-            b {{ phrase | uppercase }}
-        .col-lg-4
-        .col-lg-4
-          p {{ email }}
-      .col-sm-12.text-center
-        p Find the
-          i.fa.fa-code
-          |  of this project on
-          a(href="http://github.com/mauri870/mauri870.github.io" target="_blank")
-            i.fa.fa-github.fa-2x
-        p Built with
-          i.fa.fa-heart.heart
-          |  on top of Vue 2, Vuex, SCSS, Pug and Webpack 4
-      .col-sm-12.text-center
-        img(src="https://api.travis-ci.org/mauri870/mauri870.github.io.svg?branch=develop")
+#footerwrap
+  .container
+    .row.centered
+      .col-lg-12
+        h2
+          sup
+            i.fa.fa-quote-left
+          | {{ quote }}
+          sup
+            i.fa.fa-quote-right
+      .col-lg-4
+        p
+          b {{ phrase.toUpperCase() }}
+      .col-lg-4
+      .col-lg-4
+        p {{ email }}
+    .col-sm-12.text-center
+      p Find the
+        i.fa.fa-code
+        |  of this project on
+        a(href="http://github.com/mauri870/mauri870.github.io" target="_blank")
+          i.fa.fa-github.fa-2x
+      p Built with
+        i.fa.fa-heart.heart
+        |  on top of Vue 3, Vuex, Vue Router, SCSS, Pug and Webpack 4
+    .col-sm-12.text-center
+      img(src="https://api.travis-ci.org/mauri870/mauri870.github.io.svg?branch=develop")
 </template>
 
 <script>
