@@ -1,5 +1,5 @@
 /* eslint-disable */
-export function initAnalytics(code) {
+function initAnalytics(code) {
   (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
   (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
   m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
@@ -9,3 +9,7 @@ export function initAnalytics(code) {
   ga('send', 'pageview');
   /* eslint-enable */
 }
+
+const shuffle = (arr) => arr.sort(() => Math.random() - 0.5);
+
+export default { shuffle, initAnalytics }
